Added documentation.

+++ engine/src/core/platform/graphics/renderer/rendering_device/buffers/opengl_buffer.cpp
@@ -7,12 +7,12 @@
 
 namespace Bald::Platform::Graphics {
 
-    OpenGLBuffer::OpenGLBuffer(float* data, int count, int component_count) noexcept
+    OpenGLBuffer::OpenGLBuffer(float* data, unsigned count, unsigned component_count) noexcept
     :   m_BufferID(0),
         m_ComponentCount(component_count) {
         glGenBuffers(1, &m_BufferID);
         glBindBuffer(GL_ARRAY_BUFFER, m_BufferID);
-        glBufferData(GL_ARRAY_BUFFER, count * static_cast<int>(sizeof(float)), data, GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, count * sizeof(float), data, GL_STATIC_DRAW);
         glBindBuffer(GL_ARRAY_BUFFER, 0);
     }
